Separated migrations from schema
