+++ src/packages/dumbo/src/core/index.ts
@@ -6,6 +6,7 @@ export * from './connections';
 export * from './connectors';
 export * from './execute';
 export * from './locks';
+export * from './migrations';
 export * from './query';
 export * from './schema';
 export * from './serializer';
