+++ src/packages/dumbo/src/core/schema/index.ts
@@ -1,2 +1,1 @@
-export * from './migrations';
 export * from './schemaComponent';
